fix crash when hardlink workaround fails
we use hardlinks to workaround file permissions issues on Windows. we are unable to access the files directly because they are in use by Steam, however due to some quirk in Windows we are able to access hardlinks to those files. On some systems this does not work though and in this case SFP would crash. Let's catch this error before a crash occurs.

This might be an issue caused by users running older versions of Windows